refactor: move git_shell from core to utils
