+++ ggshield/cmd/secret/scan/prereceive.py
@@ -21,9 +21,9 @@ from ggshield.core.git_hooks.prereceive import (
     get_prereceive_timeout,
     parse_stdin,
 )
-from ggshield.core.git_shell import get_list_commit_SHA
 from ggshield.core.scan import ScanContext, ScanMode
 from ggshield.core.text_utils import display_error
+from ggshield.utils.git_shell import get_list_commit_SHA
 from ggshield.verticals.secret.output import (
     SecretGitLabWebUIOutputHandler,
     SecretOutputHandler,
