+++ ggshield/utils/git_shell.py
