+++ ggshield/verticals/secret/output/secret_text_output_handler.py
@@ -8,7 +8,6 @@ from pygitguardian.models import Match, PolicyBreak
 
 from ggshield.core.errors import UnexpectedError
 from ggshield.core.filter import censor_content, leak_dictionary_by_ignore_sha
-from ggshield.core.git_shell import Filemode
 from ggshield.core.text_utils import (
     STYLE,
     Line,
@@ -21,6 +20,7 @@ from ggshield.core.text_utils import (
     translate_validity,
 )
 from ggshield.core.utils import find_match_indices, get_lines_from_content
+from ggshield.utils.git_shell import Filemode
 
 from ..secret_scan_collection import Result, SecretScanCollection
 from .secret_output_handler import SecretOutputHandler
