+++ tests/unit/verticals/secret/output/test_json_output.py
@@ -8,8 +8,8 @@ from pytest_voluptuous import Partial, S
 from voluptuous import Optional, Required, validators
 
 from ggshield.core.filter import leak_dictionary_by_ignore_sha
-from ggshield.core.git_shell import Filemode
 from ggshield.core.scan import Commit, ScanContext, ScanMode, StringScannable
+from ggshield.utils.git_shell import Filemode
 from ggshield.verticals.secret import (
     Result,
     Results,
